#74 [Design] 인트로 배경 확정, 메인 페이지 디자인 일괄 수정 - 메인페이지 전체 좌우여백 - 메인 포스트 인원 값 설정 - 메인 포스트 예약하기 버튼 삭제 - 좋아요 버튼 슬라이드로 위치 옮김 - 헤더에 로그인 버튼 추가 - 슬라이드의 이미지 100% 차지하도록 수정
