+++ src/components/common/nav/nav.module.scss
@@ -24,6 +24,7 @@
   background-color: #fff;
   border-radius: 50%;
   opacity: 0.8;
+  z-index: 20;
 
   img {
     padding: 7px;
