+++ src/styles/scss/_variables.scss
@@ -5,7 +5,7 @@ $font-family-notoSans: 'Noto Sans KR', sans-serif;
 
 // colors
 $color-Signature-purple: #7c00de;
-$color-Signature-blue: #aaccd5;
+$color-Signature-blue: #cdddff;
 $color-font-black: #1e1e1e;
 $color-slide-default: #bebebea8;
 $color-slide-active: #ebebeb;
